[Refactor] Divide constants in each file
