+++ src/constants/alerts.js
@@ -1,0 +1,7 @@
+export const ALERT = {
+  TYPE_THOUSAND_UNIT: '로또 구입 금액을 1,000원 단위로 입력해 주세요.',
+  OVER_MAX_VALUE: '최대 구매가능 금액은 100,000원 입니다.',
+  NOT_ALL_TYPED_WINNING_INPUT: '7개의 값을 모두 입력해주세요',
+  IN_RANGE_WINNING_INPUT: '1이상 45이하의 숫자를 입력해주세요',
+  DUPLICATE_VALUE_EXIST: '중복된 값이 있습니다.',
+};
