+++ src/constants/modal.js
@@ -1,0 +1,7 @@
+export const TITLE_WITH_VALUE_MAP = new Map([
+  ['3개', 5000],
+  ['4개', 50000],
+  ['5개', 1500000],
+  ['5개 + 보너스볼', 30000000],
+  ['6개', 2000000000],
+]);
