+++ src/constants/validation.js
@@ -1,0 +1,8 @@
+export const LOTTO_VALUE = {
+  MIN_NUMBER: 1,
+  MAX_NUMBER: 45,
+  MIN_PRICE: 1000,
+  MAX_PRICE: 100000,
+  WINNIN_INPUT_LENGTH: 6,
+  MAX_LOTTO_COUNT: 6,
+};
